chromedriver tests

+++ test/index.sh
@@ -1,15 +1,18 @@
+#!/bin/bash
+
 testDir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 rootDir="$(cd "${testDir}/.." && pwd)"
-BDM_SRC_DIR="$rootDir/src"
+srcDir="$rootDir/src"
 tmpDir="$rootDir/tmp"
 
 chromedriverPath="/usr/local/bin/chromedriver"
-chromedriverVersion=$(sh $testDir/assets/mock-chromedriver97)
+chromedriverVersion=$(sh $testDir/mocks/mock-chromedriver97)
+chromedriverVersionString=$(echo "$chromedriverVersion" | cut -d ' ' -f 2)
 
-chromeStableVersion=$(sh $testDir/assets/mock-chrome)
-chromeBetaVersion=$(sh $testDir/assets/mock-chrome-beta)
-chromeDevVersion=$(sh $testDir/assets/mock-chrome-dev)
-chromeCanaryVersion=$(sh $testDir/assets/mock-chrome-canary)
+chromeStableVersion=$(sh $testDir/mocks/mock-chrome)
+chromeBetaVersion=$(sh $testDir/mocks/mock-chrome-beta)
+chromeDevVersion=$(sh $testDir/mocks/mock-chrome-dev)
+chromeCanaryVersion=$(sh $testDir/mocks/mock-chrome-canary)
 
 # linuxChromeStablePath=$(which google-chrome)
 # linuxChromeBetaPath=$(which google-chrome-beta)
@@ -47,7 +50,9 @@ esac
 # Save original chromedriver so we can restore it at the end of the test
 oneTimeSetUp() {
   if command -v $chromedriver >/dev/null; then
-    sudo mv "$chromedriverPath" "$tmpDir/chromedriver"
+    if [ -f "$chromedriverPath" ]; then
+      sudo mv "$chromedriverPath" "$tmpDir/chromedriver"
+    fi
   fi
 
   if [ $os == 'Linux' ]; then
@@ -87,14 +92,30 @@ oneTimeSetUp() {
       sudo mkdir -p "$macChromePath Canary.app/$macChromeDir"
     fi
   fi
+
+  # Modify the path to use mock curl / wget commands instead of the
+  # built in ones. That way we can control the responses and check
+  # that we are hitting the correct URLs
+  originalPath=$(echo $PATH)
+  export PATH="$testDir/mocks:$PATH"
 }
 
 oneTimeTearDown() {
   if [ "$cleanedUp" == false ]; then
     cleanedUp=true
 
+    # Unmodify path
+    PATH=$(echo $originalPath)
+
+    # Clear the contents of the mock-log-file
+    > "$testDir/mocks/mock-log-file.txt"
+
+    sudo rm -f "$tmpDir/chromedriver.zip"
+
     if [ -f "$tmpDir/chromedriver" ]; then
       sudo mv "$tmpDir/chromedriver" "$chromedriverPath"
+    else
+      sudo rm -f "$chromedriverPath"
     fi
 
     if [ $os == 'Linux' ]; then
@@ -140,49 +161,47 @@ oneTimeTearDown() {
 
 # Install mock chrome and chromedriver before each tests
 setUp() {
-  sudo cp "$testDir/assets/mock-chromedriver97" "$chromedriverPath"
+  # Clear the contents of the mock-log-file
+  > "$testDir/mocks/mock-log-file.txt"
+
+  sudo cp "$testDir/mocks/mock-chromedriver97" "$chromedriverPath"
+  sudo cp "$testDir/mocks/mock-chromedriver97.zip" "$tmpDir/chromedriver.zip"
 
   if [ $os == 'Linux' ]; then
-    sudo cp "$testDir/assets/mock-chrome" "$linuxChromeStablePath"
-    sudo cp "$testDir/assets/mock-chrome-beta" "$linuxChromeBetaPath"
-    sudo cp "$testDir/assets/mock-chrome-dev" "$linuxChromeDevPath"
+    sudo cp "$testDir/mocks/mock-chrome" "$linuxChromeStablePath"
+    sudo cp "$testDir/mocks/mock-chrome-beta" "$linuxChromeBetaPath"
+    sudo cp "$testDir/mocks/mock-chrome-dev" "$linuxChromeDevPath"
   elif [ $os == 'MacOs' ]; then
-    sudo cp "$testDir/assets/mock-chrome" "$macChromeStablePath"
-    sudo cp "$testDir/assets/mock-chrome-beta" "$macChromeBetaPath"
-    sudo cp "$testDir/assets/mock-chrome-dev" "$macChromeDevPath"
-    sudo cp "$testDir/assets/mock-chrome-canary" "$macChromeCanaryPath"
+    sudo cp "$testDir/mocks/mock-chrome" "$macChromeStablePath"
+    sudo cp "$testDir/mocks/mock-chrome-beta" "$macChromeBetaPath"
+    sudo cp "$testDir/mocks/mock-chrome-dev" "$macChromeDevPath"
+    sudo cp "$testDir/mocks/mock-chrome-canary" "$macChromeCanaryPath"
   fi
-
 }
 
 #-------------------------------------------------
 # Which Chromedriver
 #-------------------------------------------------
 test_which_chromedriver_should_ouput_dir() {
-  output=$($BDM_SRC_DIR/index.sh which chromedriver)
+  output=$($srcDir/index.sh which chromedriver)
   assertEquals "$output" "$chromedriverPath"
 }
 
 test_which_chromedriver_should_error_if_not_installed() {
   sudo rm -f "$chromedriverPath"
-  output=$($BDM_SRC_DIR/index.sh which chromedriver 2>&1)
+  output=$($srcDir/index.sh which chromedriver 2>&1)
+  exitCode=$?
+  assertEquals "$exitCode" 1
   assertContains "$output" "chromedriver is not installed"
 }
 
-test_which_chromedriver_should_error_with_exit_code_1() {
-  sudo rm -f "$chromedriverPath"
-  output=$($BDM_SRC_DIR/index.sh which chromedriver 2>&1)
-  exitCode=$?
-  assertEquals "$exitCode" 1
-}
-
 test_which_chromedriver_should_accept_stable_channel() {
-  output=$($BDM_SRC_DIR/index.sh which chromedriver=stable)
+  output=$($srcDir/index.sh which chromedriver=stable)
   assertEquals "$output" "$chromedriverPath"
 }
 
 test_which_chromedriver_should_error_for_invalid_channel() {
-  output=$($BDM_SRC_DIR/index.sh which chromedriver=invalid 2>&1)
+  output=$($srcDir/index.sh which chromedriver=invalid 2>&1)
   assertContains "$output" "Chrome supported channels"
 }
 
@@ -190,7 +209,7 @@ test_which_chromedriver_should_error_for_invalid_channel() {
 # Which Chrome
 #-------------------------------------------------
 test_which_chrome_should_output_dir() {
-  output=$($BDM_SRC_DIR/index.sh which chrome)
+  output=$($srcDir/index.sh which chrome)
 
   if [ $os == "Linux" ]; then
     assertEquals "$output" "$linuxChromeStablePath"
@@ -206,24 +225,14 @@ test_which_chrome_should_error_if_not_installed() {
     sudo rm -f "$macChromeStablePath"
   fi
 
-  output=$($BDM_SRC_DIR/index.sh which chrome 2>&1)
+  output=$($srcDir/index.sh which chrome 2>&1)
+  exitCode=$?
+  assertEquals "$exitCode" 1
   assertContains "$output" "Google Chrome is not installed"
 }
 
-test_which_chrome_should_error_with_exit_code_1() {
-  if [ $os == "Linux" ]; then
-    sudo rm -f "$linuxChromeStablePath"
-  else
-    sudo rm -f "$macChromeStablePath"
-  fi
-
-  output=$($BDM_SRC_DIR/index.sh which chrome 2>&1)
-  exitCode=$?
-  assertEquals "$exitCode" 1
-}
-
 test_which_chrome_should_accept_stable_channel() {
-  output=$($BDM_SRC_DIR/index.sh which chrome=stable)
+  output=$($srcDir/index.sh which chrome=stable)
 
   if [ $os == "Linux" ]; then
     assertEquals "$output" "$linuxChromeStablePath"
@@ -233,7 +242,7 @@ test_which_chrome_should_accept_stable_channel() {
 }
 
 test_which_chrome_should_accept_beta_channel() {
-  output=$($BDM_SRC_DIR/index.sh which chrome=beta)
+  output=$($srcDir/index.sh which chrome=beta)
 
   if [ $os == "Linux" ]; then
     assertEquals "$output" "$linuxChromeBetaPath"
@@ -243,7 +252,7 @@ test_which_chrome_should_accept_beta_channel() {
 }
 
 test_which_chrome_should_accept_dev_channel() {
-  output=$($BDM_SRC_DIR/index.sh which chrome=dev)
+  output=$($srcDir/index.sh which chrome=dev)
 
   if [ $os == "Linux" ]; then
     assertEquals "$output" "$linuxChromeDevPath"
@@ -258,17 +267,24 @@ test_which_chrome_should_accept_canary_channel() {
     startSkipping
   fi
 
-  output=$($BDM_SRC_DIR/index.sh which chrome=canary 2>/dev/null)
+  output=$($srcDir/index.sh which chrome=canary 2>/dev/null)
   assertEquals "$output" "$macChromeCanaryPath"
 }
 
 test_which_chrome_should_error_for_invalid_channel() {
-  output=$($BDM_SRC_DIR/index.sh which chrome=invalid 2>&1)
+  output=$($srcDir/index.sh which chrome=invalid 2>&1)
+  exitCode=$?
+  assertEquals "$exitCode" 1
   assertContains "$output" "Chrome supported channels"
 }
 
-test_which_chrome_should_output_verbose_logs() {
-  output=$($BDM_SRC_DIR/index.sh which chrome=beta --verbose)
+test_which_chrome_should_not_output_verbose_logs() {
+  output=$($srcDir/index.sh which chrome=beta)
+  assertNotContains "$output" "Looking for path for Google Chrome Beta"
+}
+
+test_which_chrome_should_output_verbose_logs_with_flag() {
+  output=$($srcDir/index.sh which chrome=beta --verbose)
   assertContains "$output" "Looking for path for Google Chrome Beta"
 }
 
@@ -276,30 +292,27 @@ test_which_chrome_should_output_verbose_logs() {
 # Version Chromedriver
 #-------------------------------------------------
 test_version_chromedriver_should_ouput_version() {
-  output=$($BDM_SRC_DIR/index.sh version chromedriver)
+  output=$($srcDir/index.sh version chromedriver)
   assertEquals "$output" "$chromedriverVersion"
 }
 
 test_version_chromedriver_should_error_if_not_installed() {
   sudo rm -f "$chromedriverPath"
-  output=$($BDM_SRC_DIR/index.sh version chromedriver 2>&1)
+  output=$($srcDir/index.sh version chromedriver 2>&1)
+  exitCode=$?
+  assertEquals "$exitCode" 1
   assertContains "$output" "chromedriver is not installed"
 }
 
-test_version_chromedriver_should_error_with_exit_code_1() {
-  sudo rm -f "$chromedriverPath"
-  output=$($BDM_SRC_DIR/index.sh version chromedriver 2>&1)
-  exitCode=$?
-  assertEquals "$exitCode" 1
-}
-
 test_version_chromedriver_should_accept_stable_channel() {
-  output=$($BDM_SRC_DIR/index.sh version chromedriver=stable)
+  output=$($srcDir/index.sh version chromedriver=stable)
   assertEquals "$output" "$chromedriverVersion"
 }
 
 test_version_chromedriver_should_error_for_invalid_channel() {
-  output=$($BDM_SRC_DIR/index.sh version chromedriver=invalid 2>&1)
+  output=$($srcDir/index.sh version chromedriver=invalid 2>&1)
+  exitCode=$?
+  assertEquals "$exitCode" 1
   assertContains "$output" "Chrome supported channels"
 }
 
@@ -307,7 +320,7 @@ test_version_chromedriver_should_error_for_invalid_channel() {
 # Version Chrome
 #-------------------------------------------------
 test_version_chrome_should_ouput_version() {
-  output=$($BDM_SRC_DIR/index.sh version chrome)
+  output=$($srcDir/index.sh version chrome)
   assertEquals "$output" "$chromeStableVersion"
 }
 
@@ -318,55 +331,214 @@ test_version_chrome_should_error_if_not_installed() {
     sudo rm -f "$macChromeStablePath"
   fi
 
-  output=$($BDM_SRC_DIR/index.sh version chrome 2>&1)
+  output=$($srcDir/index.sh version chrome 2>&1)
+  exitCode=$?
+  assertEquals "$exitCode" 1
   assertContains "$output" "Google Chrome is not installed"
 }
 
-test_version_chrome_should_error_with_exit_code_1() {
+test_version_chrome_should_accept_stable_channel() {
+  output=$($srcDir/index.sh version chrome=stable)
+  assertEquals "$output" "$chromeStableVersion"
+}
+
+test_version_chrome_should_accept_beta_channel() {
+  output=$($srcDir/index.sh version chrome=beta)
+  assertEquals "$output" "$chromeBetaVersion"
+}
+
+test_version_chrome_should_accept_dev_channel() {
+  output=$($srcDir/index.sh version chrome=dev)
+  assertEquals "$output" "$chromeDevVersion"
+}
+
+test_version_chrome_should_accept_canary_channel() {
+  # Linux doesn't support canary channel
+  if [ $os == "Linux" ]; then
+    startSkipping
+  fi
+
+  output=$($srcDir/index.sh version chrome=canary)
+  assertEquals "$output" "$chromeCanaryVersion"
+}
+
+test_version_chrome_should_error_for_invalid_channel() {
+  output=$($srcDir/index.sh version chrome=invalid 2>&1)
+  exitCode=$?
+  assertEquals "$exitCode" 1
+  assertContains "$output" "Chrome supported channels"
+}
+
+test_version_chrome_should_not_output_verbose_logs() {
+  output=$($srcDir/index.sh version chrome=beta --verbose)
+  assertContains "$output" "Looking for path for Google Chrome Beta"
+}
+
+test_version_chrome_should_output_verbose_logs_with_flag() {
+  output=$($srcDir/index.sh version chrome=beta --verbose)
+  assertContains "$output" "Looking for path for Google Chrome Beta"
+}
+
+#-------------------------------------------------
+# Install Chromedriver
+#-------------------------------------------------
+test_install_chromedriver_should_accept_stable_channel_and_get_version_from_chrome() {
+  output=$($srcDir/index.sh install chromedriver=stable 2>/dev/null)
+  assertContains "$output" "Installing ChromeDriver 97"
+}
+
+test_install_chromedriver_should_accept_beta_channel_and_get_version_from_chrome() {
+  output=$($srcDir/index.sh install chromedriver=beta 2>/dev/null)
+  assertContains "$output" "Installing ChromeDriver 98"
+}
+
+test_install_chromedriver_should_accept_dev_channel_and_get_version_from_chrome() {
+  output=$($srcDir/index.sh install chromedriver=dev 2>/dev/null)
+  assertContains "$output" "Installing ChromeDriver 99"
+}
+
+test_install_chromedriver_should_accept_canary_channel_and_get_version_from_chrome() {
+  # Linux doesn't support canary channel
+  if [ $os == "Linux" ]; then
+    startSkipping
+  fi
+
+  output=$($srcDir/index.sh install chromedriver=canary 2>/dev/null)
+  assertContains "$output" "Installing ChromeDriver 100"
+}
+
+test_install_chromedriver_should_error_for_invalid_channel() {
+  output=$($srcDir/index.sh install chromedriver=invalid 2>&1)
+  exitCode=$?
+  assertEquals "$exitCode" 1
+  assertContains "$output" "Chrome supported channels"
+}
+
+test_install_chromedriver_should_error_if_chrome_version_is_bad() {
+  if [ $os == 'Linux' ]; then
+    sudo cp "$testDir/mocks/mock-bad-chrome" "$linuxChromeStablePath"
+  elif [ $os == 'MacOs' ]; then
+    sudo cp "$testDir/mocks/mock-bad-chrome" "$macChromeStablePath"
+  fi
+
+  output=$($srcDir/index.sh install chromedriver 2>&1)
+  exitCode=$?
+  assertEquals "$exitCode" 1
+  assertContains "$output" "Chrome version \"Not a Number\" is not a number"
+}
+
+test_install_chromedriver_should_error_if_not_installed() {
   if [ $os == "Linux" ]; then
     sudo rm -f "$linuxChromeStablePath"
   else
     sudo rm -f "$macChromeStablePath"
   fi
 
-  output=$($BDM_SRC_DIR/index.sh version chrome 2>&1)
+  output=$($srcDir/index.sh install chromedriver=stable 2>&1)
   exitCode=$?
   assertEquals "$exitCode" 1
+  assertContains "$output" "Google Chrome is not installed"
 }
 
-test_version_chrome_should_accept_stable_channel() {
-  output=$($BDM_SRC_DIR/index.sh version chrome=stable)
-  assertEquals "$output" "$chromeStableVersion"
+
+test_install_chromedriver_should_print_chrome_version() {
+  output=$($srcDir/index.sh install chromedriver=stable 2>/dev/null)
+  assertContains "$output" "Chrome Stable version detected as 97"
 }
 
-test_version_chrome_should_accept_beta_channel() {
-  output=$($BDM_SRC_DIR/index.sh version chrome=beta)
-  assertEquals "$output" "$chromeBetaVersion"
+test_install_chromedriver_should_print_chromedriver_version_to_install() {
+  output=$($srcDir/index.sh install chromedriver=97 2>/dev/null)
+  assertContains "$output" "Installing ChromeDriver 97"
 }
 
-test_version_chrome_should_accept_dev_channel() {
-  output=$($BDM_SRC_DIR/index.sh version chrome=dev)
-  assertEquals "$output" "$chromeDevVersion"
+test_install_chromedriver_should_get_lastest_release_using_version() {
+  $srcDir/index.sh install chromedriver=97 >/dev/null 2>/dev/null
+  mockLogs=$(cat "$testDir/mocks/mock-log-file.txt")
+  assertContains "$mockLogs" "https://chromedriver.storage.googleapis.com/LATEST_RELEASE_97"
 }
 
-test_version_chrome_should_accept_stable_channel() {
-  # Linux doesn't support canary channel
+test_install_chromedriver_should_use_one_version_lower_and_try_again_if_version_is_not_found() {
+  # version >100 is used in mock-http to return a bad request
+  output=$($srcDir/index.sh install chromedriver=101 2>/dev/null)
+  mockLogs=$(cat "$testDir/mocks/mock-log-file.txt")
+  assertContains "$output" "ChromeDriver 101 not found. Retrying with ChromeDriver 100"
+  assertContains "$mockLogs" "https://chromedriver.storage.googleapis.com/LATEST_RELEASE_101"
+  assertContains "$mockLogs" "https://chromedriver.storage.googleapis.com/LATEST_RELEASE_100"
+}
+
+test_install_chromedriver_should_error_if_chromedriver_version_is_not_available() {
+  output=$($srcDir/index.sh install chromedriver=200 2>&1)
+  exitCode=$?
+  assertEquals "$exitCode" 1
+  assertContains "$output" "Unable to get ChromeDriver version; Something went wrong"
+}
+
+test_install_chromedriver_should_exit_if_chromedriver_version_already_is_installed() {
+  output=$($srcDir/index.sh install chromedriver=97 2>/dev/null)
+  exitCode=$?
+  assertEquals "$exitCode" 0
+  assertContains "$output" "ChromeDriver $chromedriverVersionString already installed"
+}
+
+test_install_chromedriver_should_download_zip_from_version() {
+  sudo rm -f "$chromedriverPath"
+  output=$($srcDir/index.sh install chromedriver=97 2>/dev/null)
+  mockLogs=$(cat "$testDir/mocks/mock-log-file.txt")
+
   if [ $os == "Linux" ]; then
-    startSkipping
+    assertContains "$mockLogs" "https://chromedriver.storage.googleapis.com/$chromedriverVersionString/chromedriver_linux64.zip"
+  else
+    assertContains "$mockLogs" "https://chromedriver.storage.googleapis.com/$chromedriverVersionString/chromedriver_mac64.zip"
   fi
-
-  output=$($BDM_SRC_DIR/index.sh version chrome=canary)
-  assertEquals "$output" "$chromeCanaryVersion"
 }
 
-test_version_chrome_should_error_for_invalid_channel() {
-  output=$($BDM_SRC_DIR/index.sh version chrome=invalid 2>&1)
-  assertContains "$output" "Chrome supported channels"
+test_install_chromedriver_should_unzip_download() {
+  sudo rm -f "$chromedriverPath"
+  $srcDir/index.sh install chromedriver=97 >/dev/null 2>/dev/null
+  mockLogs=$(cat "$testDir/mocks/mock-log-file.txt")
+  assertContains "$mockLogs" "unzip $tmpDir/chromedriver.zip -d $tmpDir"
 }
 
-test_version_chrome_should_output_verbose_logs() {
-  output=$($BDM_SRC_DIR/index.sh version chrome=beta --verbose)
-  assertContains "$output" "Looking for path for Google Chrome Beta"
+test_install_chromedriver_should_change_permissions_of_download() {
+  sudo rm -f "$chromedriverPath"
+  $srcDir/index.sh install chromedriver=97 >/dev/null 2>/dev/null
+  mockLogs=$(cat "$testDir/mocks/mock-log-file.txt")
+  assertContains "$mockLogs" "chmod +x $tmpDir/chromedriver"
+}
+
+test_install_chromedriver_should_move_download() {
+  sudo rm -f "$chromedriverPath"
+  $srcDir/index.sh install chromedriver=97 >/dev/null 2>/dev/null
+  mockLogs=$(cat "$testDir/mocks/mock-log-file.txt")
+  assertContains "$mockLogs" "mv $tmpDir/chromedriver /usr/local/bin"
+}
+
+test_install_chromedriver_should_cleanup_zip() {
+  sudo rm -f "$chromedriverPath"
+  $srcDir/index.sh install chromedriver=97 >/dev/null 2>/dev/null
+
+  if [[ -f "$tmpDir/chromedriver.zip" ]]; then
+    fail "Zip was not removed"
+  fi
+}
+
+test_install_chromedriver_should_verify_install_matches_version() {
+  sudo rm -f "$chromedriverPath"
+  output=$($srcDir/index.sh install chromedriver=97 2>/dev/null)
+  assertContains "$output" "Successfully installed ChromeDriver $chromedriverVersionString"
+}
+
+test_install_chromedriver_should_error_if_installed_version_does_not_match() {
+  sudo rm -f "$chromedriverPath"
+  output=$($srcDir/index.sh install chromedriver=98 2>&1)
+  exitCode=$?
+  assertEquals "$exitCode" 1
+  assertContains "$output" "Unable to install ChromeDriver; Something went wrong"
+}
+
+test_install_chromedriver_should_output_verbose_logs_with_flag() {
+  output=$($srcDir/index.sh install chromedriver=97 --verbose)
+  assertContains "$output" "Received response of $chromedriverVersionString"
 }
 
 # Load shUnit2
